Rewrite and got the custom entity replacing the EntityFox
I found out how to get my custom entity to replace the vanilla EntityFox. This should fix most of the errors people were getting.

I also rewrote the entire plugin. It saves some data with SQLite and some parts of it are cleaned up, optimized, or certain things moved in an appropriate class.